backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.3
motor==3.3.2
pydantic==2.4.2
pydantic[email]==2.4.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
python-dotenv==1.0.0
passlib[bcrypt]
